+++ public/index.html
@@ -20,6 +20,6 @@
   </div>
 </body>
 
-<script type="module" src="../src/index.js"></script>
+<script type="module" src="src/index.js"></script>
 
 </html>
